Thêm api đặt trước một hoặc nhiều sách, xem danh sách đặt trước của một user cụ thể

+++ src/main/java/com/library/libraryapi/controllers/BookController.java
@@ -23,7 +23,7 @@ public class BookController {
     private BookService bookService;
 
     @GetMapping("/{bookId}")
-    public ResponseEntity<?> getBookById(@PathVariable Long bookId) {
+    public ResponseEntity<?> getBookById(@PathVariable Integer bookId) {
         Optional<Book> book = bookService.getBookById(bookId);
         if (book.isPresent()) {
             return ResponseEntity.ok(book.get());
