+++ src/main/java/com/library/libraryapi/repository/UsersRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.library.libraryapi.models.Users;
 
 @Repository
-public interface UsersRepository extends JpaRepository<Users, Long>{
+public interface UsersRepository extends JpaRepository<Users, Integer>{
 	Optional<Users> findByUsername(String username);
     Optional<Users> findByEmail(String email);
 }
