+++ src/main/java/com/library/libraryapi/services/BookService.java
@@ -14,7 +14,7 @@ public class BookService {
     @Autowired
     private BookRepository bookRepository;
 
-    public Optional<Book> getBookById(Long bookId) {
+    public Optional<Book> getBookById(Integer bookId) {
         return bookRepository.findById(bookId);
     }
 }
